In this version can be possible to run program by args from cmd, args and data from file and from api alphavantage.
