+++ InputData/INTC.csv
@@ -6312,7 +6312,7 @@
 01/07/2017,32.650002
 02/07/2017,32.650002
 03/07/2017,32.650002
-03/07/2017,34.02
+04/07/2017,34.02
 05/07/2017,33.77
 06/07/2017,33.889999
 07/07/2017,34.189999
