--- sample raw data.csv
@@ -1,12 +1,0 @@
-02/01/2000,72.220993
-03/01/2000,72.220993
-04/01/2000,72.555634
-05/01/2000,76.738297
-06/01/2000,75.734467
-07/01/2000,75.511398
-08/01/2000,75.511398
-09/01/2000,75.511398
-10/01/2000,75.288315
-10/02/2000,75.288315
-11/02/2000,76.288315
-03/03/2000,75.288315
